refactor(components): refactor suggestionApi service
более правильный вариант сервиса подсказок

МЕТА ДАННЫЕ: refactor 06
